Add CI support through Docker (#32)
* Stub in Docker support, not working yet.

Fix spelling mistake in CMakeLists.txt also

* Add Docker support,`fix spelling in CMakeLists.txt

Running `scripts/docker-build.sh` will build genus in a Docker container

* Adding first cut of Jenkinsfile

* Fix script to run in Jenkinsfile

* Tweak strategy for creative-engine directory

This is ugly but we need to be smart enough to do the
right thing both when run on a local development workstation
and from Jenkins.

* Ensure clean build

* Fix tabs -> spaces

* See if we can omit checkout scm

* Seeeparate stages further

+++ scripts/common.sh
@@ -45,10 +45,15 @@ function ensure_creative_engine {
 
 function build {
     cd "$BASE_DIR" || exit 1
-    rm -f creative-engine
-    ln -s ../creative-engine . 
+    if [[ ! -d creative-engine ]]; then
+        rm -f creative-engine
+        ln -s ../creative-engine . 
+    fi
     mkdir -p "$BUILD_DIR"
     cd "$BUILD_DIR" || exit 1
+    pwd
+    ls -l
+    ls -l ..
     cmake ..
     make
 }
